requirements.txt
affine==2.4.0
alabaster==0.7.13
ansi2html==1.9.1
antlr4-python3-runtime==4.9.3
anyio==4.1.0
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
asciitree==0.3.3
asttokens==2.4.1
async-lru==2.0.4
attrs==23.1.0
Babel==2.13.1
beautifulsoup4==4.12.2
bleach==6.1.0
blinker==1.7.0
bokeh==3.3.2
catboost==1.2.2
certifi==2023.11.17
cffi==1.16.0
cfgv==3.4.0
charset-normalizer==3.3.2
click==8.1.7
click-plugins==1.1.1
cligj==0.7.2
cloudpickle==3.0.0
colorama==0.4.6
coloredlogs==15.0.1
comm==0.2.0
contourpy==1.2.0
cycler==0.12.1
dash==2.14.2
dash-bootstrap-components==1.5.0
dash-bootstrap-templates==1.1.1
dash-core-components==2.0.0
dash-html-components==2.0.0
dash-table==5.0.0
dask==2023.12.0
dask-glm==0.3.2
dask-image==2023.8.1
dask-ml==2023.3.24
debugpy==1.8.0
decorator==5.1.1
defusedxml==0.7.1
distlib==0.3.8
distributed==2023.12.0
docutils==0.20.1
efficientnet-pytorch==0.7.1
exceptiongroup==1.2.0
executing==2.0.1
fasteners==0.19
fastjsonschema==2.19.0
filelock==3.13.1
flake8==6.1.0
Flask==3.0.0
fonttools==4.46.0
fqdn==1.5.1
fsspec==2023.5.0
graphviz==0.20.1
hiddenlayer==0.3
huggingface-hub==0.19.4
humanfriendly==10.0
hydra-core==1.3.2
identify==2.5.33
idna==3.6
imagecodecs==2023.9.18
imageio==2.33.0
imagesize==1.4.1
importlib-metadata==7.0.0
ipykernel==6.27.1
ipython==8.18.1
ipywidgets==8.1.1
isoduration==20.11.0
itsdangerous==2.1.2
jedi==0.19.1
Jinja2==3.1.2
joblib==1.3.2
json5==0.9.14
jsonpointer==2.4
jsonschema==4.20.0
jsonschema-specifications==2023.11.2
jupyter==1.0.0
jupyter-console==6.6.3
jupyter-events==0.9.0
jupyter-lsp==2.2.1
jupyter_client==8.6.0
jupyter_core==5.5.0
jupyter_server==2.12.1
jupyter_server_terminals==0.4.4
jupyterlab==4.0.9
jupyterlab-widgets==3.0.9
jupyterlab_pygments==0.3.0
jupyterlab_server==2.25.2
kiwisolver==1.4.5
lazy_loader==0.3
llvmlite==0.41.1
locket==1.0.0
markdown-it-py==3.0.0
MarkupSafe==2.1.3
matplotlib==3.8.2
matplotlib-inline==0.1.6
mccabe==0.7.0
mdit-py-plugins==0.4.0
mdurl==0.1.2
mistune==3.0.2
mpmath==1.3.0
msgpack==1.0.7
multipledispatch==1.0.0
munch==4.0.0
mypy==1.7.1
mypy-extensions==1.0.0
myst-parser==2.0.0
nbclient==0.9.0
nbconvert==7.12.0
nbformat==5.9.2
nest-asyncio==1.5.8
networkx==3.2.1
nodeenv==1.8.0
notebook==7.0.6
notebook_shim==0.2.3
numba==0.58.1
numcodecs==0.12.1
numpy==1.26.2
omegaconf==2.3.0
opencv-python==4.9.0.80
overrides==7.4.0
packaging==23.2
pandas==2.1.4
pandas-stubs==2.1.1.230928
pandocfilters==1.5.0
parso==0.8.3
partd==1.4.1
Pillow==10.1.0
PIMS==0.6.1
platformdirs==4.1.0
plotly==5.18.0
polars==0.19.19
pre-commit==3.6.0
pretrainedmodels==0.7.4
prometheus-client==0.19.0
prompt-toolkit==3.0.41
psutil==5.9.6
pure-eval==0.2.2
pycodestyle==2.11.1
pycparser==2.21
pydocstyle==6.3.0
pyflakes==3.1.0
pygit2==1.13.3
Pygments==2.17.2
pyngrok==7.0.3
pynvml==11.4.1
pyparsing==3.1.1
pyreadline3==3.4.1
python-dateutil==2.8.2
python-json-logger==2.0.7
pytz==2023.3.post1
pywin32==306
pywinpty==2.0.12
PyYAML==6.0.1
pyzmq==25.1.2
qtconsole==5.5.1
QtPy==2.4.1
rasterio==1.3.9
referencing==0.32.0
requests==2.31.0
retrying==1.3.4
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rpds-py==0.13.2
safetensors==0.4.1
scikit-image==0.22.0
scikit-learn==1.4.0
scipy==1.11.4
seaborn==0.13.0
segmentation-models-pytorch==0.3.3
Send2Trash==1.8.2
six==1.16.0
skorch==0.15.0
slicerator==1.1.0
sniffio==1.3.0
snowballstemmer==2.2.0
snuggs==1.4.7
sortedcontainers==2.4.0
soupsieve==2.5
sparse==0.14.0
Sphinx==7.2.6
sphinx-autodoc-typehints==1.25.2
sphinxawesome-theme==5.0.0
sphinxcontrib-applehelp==1.0.7
sphinxcontrib-devhelp==1.0.5
sphinxcontrib-htmlhelp==2.0.4
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.6
sphinxcontrib-serializinghtml==1.1.9
stack-data==0.6.3
sympy==1.12
tabulate==0.9.0
tblib==3.0.0
tenacity==8.2.3
terminado==0.18.0
threadpoolctl==3.2.0
tifffile==2023.9.26
timm==0.9.2
tinycss2==1.2.1
tomli==2.0.1
toolz==0.12.0
--find-links https://download.pytorch.org/whl/torch_stable.html
torch==2.1.2+cu118
torchaudio==2.1.2+cu118
torchvision==0.16.2+cu118
torch-summary==1.4.5
torchviz==0.0.2
tornado==6.4
tqdm==4.66.1
traitlets==5.14.1
types-python-dateutil==2.8.19.14
types-pytz==2023.3.1.1
typing_extensions==4.9.0
tzdata==2023.3
uri-template==1.3.0
urllib3==2.1.0
virtualenv==20.25.0
wcwidth==0.2.12
webcolors==1.13
webencodings==0.5.1
websocket-client==1.7.0
Werkzeug==3.0.1
widgetsnbextension==4.0.9
xyzservices==2023.10.1
zarr==2.16.1
albumentations==1.3.1
zict==3.0.0
zipp==3.17.0
wandb==0.16.2
annotated-types==0.6.0
randomname==0.2.1
open-clip-torch==2.23.0
sentence-transformers==2.2.2
monai==1.3.0
einops==0.7.0
kornia==0.7.1
kornia-rs==0.0.8
xgboost==2.0.3
lightgbm==4.2.0
MedPy==0.4.0
-e ./
